refactor(nomenclatura); Refatorado e corrigindo estrutura dos nomes e e organização da classe
