+++ desafio-casdastro/src/main/java/br/desafio/prodiga/config/SpringFoxConfig.java
@@ -1,4 +1,4 @@
-package br.desafio.prodiga.Config;
+package br.desafio.prodiga.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
