+++ desafio-casdastro/src/main/java/br/desafio/prodiga/controller/ClienteController.java
@@ -1,4 +1,4 @@
-package br.desafio.prodiga.Controller;
+package br.desafio.prodiga.controller;
 
 import java.util.List;
 
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import br.desafio.prodiga.Model.Cliente;
-import br.desafio.prodiga.Service.ClienteService;
+import br.desafio.prodiga.model.Cliente;
+import br.desafio.prodiga.service.ClienteService;
 
 @RestController
 @RequestMapping("/api/clientes")
