+++ desafio-casdastro/src/main/java/br/desafio/prodiga/enums/SituacaoFatura.java
@@ -1,4 +1,4 @@
-package br.desafio.prodiga.Enums;
+package br.desafio.prodiga.enums;
 
 public enum SituacaoFatura {
   GERADA, PAGA, CANCELADA
