+++ desafio-casdastro/src/main/java/br/desafio/prodiga/model/Cliente.java
@@ -1,4 +1,4 @@
-package br.desafio.prodiga.Model;
+package br.desafio.prodiga.model;
