+++ desafio-casdastro/src/main/java/br/desafio/prodiga/repository/ClienteRepository.java
@@ -1,8 +1,8 @@
-package br.desafio.prodiga.Repository;
+package br.desafio.prodiga.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import br.desafio.prodiga.Model.Cliente;
+import br.desafio.prodiga.model.Cliente;
 
 @Repository
 public interface ClienteRepository extends JpaRepository<Cliente, Long>{
